GERALT IMPORTANT CHANGES
Upgraded GeraltScript_v3, halfway done with animation timing and polishing, I still need to do the combo frame times and some other minor stuff.

There was a very minor bug on the calculation of the magnitude of the direction vector on GeraltScript_v1 and the P1 and P2 versions of the Setup Scene Vertical_Slice_v2, now fixed.