New Engine build v0.4.0.5
# CHANGELOG v0.4.0.5
### General Notes
When putting a scene, make it static!!! It will improve performance! You can put an object to static and it will ask if all child objects have to be static too.

Old scenes won't work, so, make sure you reload all

All FBX should be exported from other programs (as max) in **meters**

Is your engine now :)

### Engine Functionalities
* Gamma Correction Implementation (Rendering window on Window --> Rendering)
* Materials preview in the Project Panel, in assets
* Framebuffer is rendered now in the Game Build, making it to look as the engine's scene
* Issues on textures in materials fixed (they sometimes become nullptr, they don't now :) )
* Mesh Renderer Component Active/Unactive checkbox works now
* Rendering pipeline re-organized to make it better and a bit faster (in terms of uniforms and readablity)

### Prefabs Basic Implementation
They can now be created and will work as expected, except for modifying them once in scene. Once you put them on scene, they will no longer be overwritten if the original is. We are working on it :)